recommended code analysis changes

+++ server/src/main/java/dataaccess/MemoryAuth.java
@@ -6,20 +6,26 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
 
-public class MemoryAuth implements AuthDAO{
+public class MemoryAuth implements AuthDAO {
     private final HashMap<String, AuthData> db = new HashMap<>();
 
     public MemoryAuth() {
     }
 
     public void addAuth(AuthData auth) throws DataAccessException {
+        if (null == auth || null == auth.username() || null == auth.authToken()) {
+            throw new DataAccessException("authData is malformed");
+        }
         if (getAuthByAuthToken(auth.authToken()) != null) {
             throw new DataAccessException("AuthToken already in database");
         }
         db.put(auth.authToken(), auth);
     }
 
-    public AuthData getAuthByUsername(String username) {
+    public AuthData getAuthByUsername(String username) throws DataAccessException {
+        if (null == username) {
+            throw new DataAccessException("user is null");
+        }
         for (Map.Entry<String, AuthData> entry : db.entrySet()) {
             if (Objects.equals(entry.getValue().username(), username)) {
                 return entry.getValue();
@@ -28,11 +34,17 @@ public class MemoryAuth implements AuthDAO{
         return null;
     }
 
-    public AuthData getAuthByAuthToken(String authToken) {
+    public AuthData getAuthByAuthToken(String authToken) throws DataAccessException {
+        if (null == authToken) {
+            throw new DataAccessException("authToken is null");
+        }
         return db.get(authToken);
     }
 
-    public void deleteAuthByAuthToken(String authToken) {
+    public void deleteAuthByAuthToken(String authToken) throws DataAccessException {
+        if (null == authToken) {
+            throw new DataAccessException("authToken is null");
+        }
         db.remove(authToken);
     }
